# src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: Car Rental API
  description: REST API for car rental service
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      description: Returns a user by their ID
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User successfully retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Users
      summary: Update user
      description: Updates an existing user
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
        required: true
      responses:
        "200":
          description: User successfully updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid user data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Users
      summary: Delete user
      description: Deletes a user
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: User successfully deleted
        "403":
          description: Access denied
        "404":
          description: User not found
      security:
      - bearerAuth: []
...
  /api/v1/cars/{id}:
    get:
      tags:
      - Cars
      summary: Get car by ID
      description: Returns an active car by its ID
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Car successfully retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
        "404":
          description: Car not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Cars
      summary: Update car
      description: Updates an existing car. Available only to administrators
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdateRequest"
        required: true
      responses:
        "200":
          description: Car successfully updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
        "400":
          description: Invalid car data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
        "404":
          description: Car not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
        "409":
          description: Car with this license plate already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Cars
      summary: Delete car
      description: Deletes a car. Available only to administrators
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Car successfully deleted
        "403":
          description: Access denied
        "404":
          description: Car not found
        "409":
          description: Car cannot be deleted because it is in use
      security:
      - bearerAuth: []
  /api/v1/car-categories/{id}:
    get:
      tags:
      - Car Categories
      summary: Get car category by ID
      description: Returns a car category by its ID
      operationId: getById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Category successfully retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Car Categories
      summary: Update car category
      description: Updates an existing car category. Available only to administrators
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarCategoryUpdateRequest"
        required: true
      responses:
        "200":
          description: Category successfully updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "400":
          description: Invalid category data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "404":
          description: Category not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Car Categories
      summary: Delete car category
      description: Deletes a car category. Available only to administrators
      operationId: delete_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Category successfully deleted
        "403":
          description: Access denied
        "404":
          description: Category not found
        "409":
          description: Category is in use by cars
      security:
      - bearerAuth: []
...
  /api/v1/car-categories:
    get:
      tags:
      - Car Categories
      summary: Get all car categories
      description: Returns all car categories
      operationId: getAll_2
      responses:
        "200":
          description: Categories successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarCategoryResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Car Categories
      summary: Create car category
      description: Creates a new car category. Available only to administrators
      operationId: create_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarCategoryCreateRequest"
        required: true
      responses:
        "201":
          description: Category successfully created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "400":
          description: Invalid category data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CarCategoryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Get all users
      description: Returns all users. Available only to administrators
      operationId: getAll_4
      responses:
        "200":
          description: Users successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
  /api/v1/users/username/{username}:
    get:
      tags:
      - Users
      summary: Get user by username
      description: Returns a user by their username
      operationId: getByUsername
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: User successfully retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
...
  /api/v1/reviews/car/{carId}:
    get:
      tags:
      - Reviews
      summary: Get reviews by car ID
      description: Returns all reviews for the specified car
      operationId: getByCarId
      parameters:
      - name: carId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Reviews successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
        "404":
          description: Car not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
      security:
      - bearerAuth: []
...
  /api/v1/bookings/user/{userId}:
    get:
      tags:
      - Bookings
      summary: Get bookings by user id
      description: Returns all user bookings. Available only to administrators
      operationId: getByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Bookings successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
    UserUpdateRequest:
      type: object
      properties:
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        active:
          type: boolean
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CarUpdateRequest:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        productionYear:
          type: integer
          format: int32
        licensePlate:
          type: string
        color:
          type: string
        transmission:
          type: string
          enum:
          - AUTOMATIC
          - MANUAL
        fuelType:
          type: string
          enum:
          - PETROL
          - DIESEL
          - ELECTRIC
          - HYBRID
        seats:
          type: integer
          format: int32
          minimum: 1
        active:
          type: boolean
        categoryId:
          type: integer
          format: int64
    CarResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        model:
          type: string
        productionYear:
          type: integer
          format: int32
        licensePlate:
          type: string
        color:
          type: string
        transmission:
          type: string
          enum:
          - AUTOMATIC
          - MANUAL
        fuelType:
          type: string
          enum:
          - PETROL
          - DIESEL
          - ELECTRIC
          - HYBRID
        seats:
          type: integer
          format: int32
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CarCategoryUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
          minimum: 0.01
    CarCategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        carId:
          type: integer
          format: int64
        carBrand:
          type: string
        carModel:
          type: string
        bookingId:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CarCategoryCreateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        pricePerDay:
          type: number
          minimum: 0.01
      required:
      - name
      - pricePerDay
...
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        carBrand:
          type: string
        carModel:
          type: string
        pickupLocationId:
          type: integer
          format: int64
        pickupLocationName:
          type: string
        returnLocationId:
          type: integer
          format: int64
        returnLocationName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalPrice:
          type: number
        status:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - CANCELLED
          - COMPLETED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT